requirements.txt
numpy== 1.26.4
pandas== 2.2.3
matplotlib== 3.7.2
seaborn== 0.12.2
torch== 2.6.0+cu124
torchvision== 0.21.0+cu124
scikit-learn== 1.2.2
tqdm== 4.67.1
Pillow== 11.2.1
opencv-python== 4.11.0